Add filter dropdown for single entity
