+++ front/src/modules/ui/filter-n-sort/components/FilterDropdownButton.tsx
@@ -1,25 +1,12 @@
-import { Context, useCallback, useState } from 'react';
+import { Context } from 'react';
 
-import { DropdownMenuSeparator } from '@/ui/dropdown/components/DropdownMenuSeparator';
-import { filterDefinitionUsedInDropdownScopedState } from '@/ui/filter-n-sort/states/filterDefinitionUsedInDropdownScopedState';
-import { filterDropdownSearchInputScopedState } from '@/ui/filter-n-sort/states/filterDropdownSearchInputScopedState';
-import { filtersScopedState } from '@/ui/filter-n-sort/states/filtersScopedState';
-import { isFilterDropdownOperandSelectUnfoldedScopedState } from '@/ui/filter-n-sort/states/isFilterDropdownOperandSelectUnfoldedScopedState';
-import { selectedOperandInDropdownScopedState } from '@/ui/filter-n-sort/states/selectedOperandInDropdownScopedState';
-import { useSetHotkeyScope } from '@/ui/utilities/hotkey/hooks/useSetHotkeyScope';
 import { useRecoilScopedState } from '@/ui/utilities/recoil-scope/hooks/useRecoilScopedState';
 
+import { availableFiltersScopedState } from '../states/availableFiltersScopedState';
 import { FiltersHotkeyScope } from '../types/FiltersHotkeyScope';
 
-import DropdownButton from './DropdownButton';
-import { FilterDropdownDateSearchInput } from './FilterDropdownDateSearchInput';
-import { FilterDropdownEntitySearchInput } from './FilterDropdownEntitySearchInput';
-import { FilterDropdownEntitySelect } from './FilterDropdownEntitySelect';
-import { FilterDropdownFilterSelect } from './FilterDropdownFilterSelect';
-import { FilterDropdownNumberSearchInput } from './FilterDropdownNumberSearchInput';
-import { FilterDropdownOperandButton } from './FilterDropdownOperandButton';
-import { FilterDropdownOperandSelect } from './FilterDropdownOperandSelect';
-import { FilterDropdownTextSearchInput } from './FilterDropdownTextSearchInput';
+import { MultipleFiltersDropdownButton } from './MultipleFiltersDropdownButton';
+import { SingleEntityFilterDropdownButton } from './SingleEntityFilterDropdownButton';
 
 export function FilterDropdownButton({
   context,
@@ -28,93 +15,20 @@ export function FilterDropdownButton({
   context: Context<string | null>;
   HotkeyScope: FiltersHotkeyScope;
 }) {
-  const [isUnfolded, setIsUnfolded] = useState(false);
-
-  const [
-    isFilterDropdownOperandSelectUnfolded,
-    setIsFilterDropdownOperandSelectUnfolded,
-  ] = useRecoilScopedState(
-    isFilterDropdownOperandSelectUnfoldedScopedState,
+  const [availableFilters] = useRecoilScopedState(
+    availableFiltersScopedState,
     context,
   );
-
-  const [filterDefinitionUsedInDropdown, setFilterDefinitionUsedInDropdown] =
-    useRecoilScopedState(filterDefinitionUsedInDropdownScopedState, context);
-
-  const [, setFilterDropdownSearchInput] = useRecoilScopedState(
-    filterDropdownSearchInputScopedState,
-    context,
-  );
-
-  const [filters] = useRecoilScopedState(filtersScopedState, context);
-
-  const [selectedOperandInDropdown, setSelectedOperandInDropdown] =
-    useRecoilScopedState(selectedOperandInDropdownScopedState, context);
-
-  const resetState = useCallback(() => {
-    setIsFilterDropdownOperandSelectUnfolded(false);
-    setFilterDefinitionUsedInDropdown(null);
-    setSelectedOperandInDropdown(null);
-    setFilterDropdownSearchInput('');
-  }, [
-    setFilterDefinitionUsedInDropdown,
-    setSelectedOperandInDropdown,
-    setFilterDropdownSearchInput,
-    setIsFilterDropdownOperandSelectUnfolded,
-  ]);
-
-  const isFilterSelected = (filters?.length ?? 0) > 0;
-
-  const setHotkeyScope = useSetHotkeyScope();
-
-  function handleIsUnfoldedChange(newIsUnfolded: boolean) {
-    if (newIsUnfolded) {
-      setHotkeyScope(HotkeyScope);
-      setIsUnfolded(true);
-    } else {
-      if (filterDefinitionUsedInDropdown?.type === 'entity') {
-        setHotkeyScope(HotkeyScope);
-      }
-      setIsUnfolded(false);
-      resetState();
-    }
-  }
-
-  return (
-    <DropdownButton
-      label="Filter"
-      isActive={isFilterSelected}
-      isUnfolded={isUnfolded}
-      onIsUnfoldedChange={handleIsUnfoldedChange}
+  console.log('availableFilters from generic', availableFilters);
+  return availableFilters.length > 1 ? (
+    <MultipleFiltersDropdownButton
+      context={context}
       HotkeyScope={HotkeyScope}
-    >
-      {!filterDefinitionUsedInDropdown ? (
-        <FilterDropdownFilterSelect context={context} />
-      ) : isFilterDropdownOperandSelectUnfolded ? (
-        <FilterDropdownOperandSelect context={context} />
-      ) : (
-        selectedOperandInDropdown && (
-          <>
-            <FilterDropdownOperandButton context={context} />
-            <DropdownMenuSeparator />
-            {filterDefinitionUsedInDropdown.type === 'text' && (
-              <FilterDropdownTextSearchInput context={context} />
-            )}
-            {filterDefinitionUsedInDropdown.type === 'number' && (
-              <FilterDropdownNumberSearchInput context={context} />
-            )}
-            {filterDefinitionUsedInDropdown.type === 'date' && (
-              <FilterDropdownDateSearchInput context={context} />
-            )}
-            {filterDefinitionUsedInDropdown.type === 'entity' && (
-              <FilterDropdownEntitySearchInput context={context} />
-            )}
-            {filterDefinitionUsedInDropdown.type === 'entity' && (
-              <FilterDropdownEntitySelect context={context} />
-            )}
-          </>
-        )
-      )}
-    </DropdownButton>
+    />
+  ) : (
+    <SingleEntityFilterDropdownButton
+      context={context}
+      HotkeyScope={HotkeyScope}
+    />
   );
 }
