+++ front/src/modules/ui/filter-n-sort/components/SingleEntityFilterDropdownButton.tsx
@@ -1,0 +1,120 @@
+import { Context, useState } from 'react';
+import React from 'react';
+import styled from '@emotion/styled';
+
+import { filterDefinitionUsedInDropdownScopedState } from '@/ui/filter-n-sort/states/filterDefinitionUsedInDropdownScopedState';
+import { filterDropdownSearchInputScopedState } from '@/ui/filter-n-sort/states/filterDropdownSearchInputScopedState';
+import { selectedOperandInDropdownScopedState } from '@/ui/filter-n-sort/states/selectedOperandInDropdownScopedState';
+import { useSetHotkeyScope } from '@/ui/utilities/hotkey/hooks/useSetHotkeyScope';
+import { useRecoilScopedState } from '@/ui/utilities/recoil-scope/hooks/useRecoilScopedState';
+
+import { availableFiltersScopedState } from '../states/availableFiltersScopedState';
+import { filtersScopedState } from '../states/filtersScopedState';
+import { FiltersHotkeyScope } from '../types/FiltersHotkeyScope';
+import { getOperandsForFilterType } from '../utils/getOperandsForFilterType';
+
+import { DropdownMenuContainer } from './DropdownMenuContainer';
+import { FilterDropdownEntitySearchInput } from './FilterDropdownEntitySearchInput';
+import { FilterDropdownEntitySelect } from './FilterDropdownEntitySelect';
+
+const StyledDropdownButtonContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  z-index: 1;
+`;
+
+type StyledDropdownButtonProps = {
+  isUnfolded: boolean;
+};
+
+const StyledDropdownButton = styled.div<StyledDropdownButtonProps>`
+  background: ${({ theme }) => theme.background.primary};
+  border-radius: ${({ theme }) => theme.border.radius.sm};
+  cursor: pointer;
+  display: flex;
+  filter: ${(props) => (props.isUnfolded ? 'brightness(0.95)' : 'none')};
+
+  padding: ${({ theme }) => theme.spacing(1)};
+  padding-left: ${({ theme }) => theme.spacing(2)};
+
+  padding-right: ${({ theme }) => theme.spacing(2)};
+  user-select: none;
+
+  &:hover {
+    filter: brightness(0.95);
+  }
+`;
+
+export function SingleEntityFilterDropdownButton({
+  context,
+  HotkeyScope,
+}: {
+  context: Context<string | null>;
+  HotkeyScope: FiltersHotkeyScope;
+}) {
+  const [availableFilters] = useRecoilScopedState(
+    availableFiltersScopedState,
+    context,
+  );
+  const availableFilter = availableFilters[0];
+
+  const [isUnfolded, setIsUnfolded] = useState(false);
+
+  const [filters] = useRecoilScopedState(filtersScopedState, context);
+
+  const [filterDefinitionUsedInDropdown, setFilterDefinitionUsedInDropdown] =
+    useRecoilScopedState(filterDefinitionUsedInDropdownScopedState, context);
+
+  const [, setFilterDropdownSearchInput] = useRecoilScopedState(
+    filterDropdownSearchInputScopedState,
+    context,
+  );
+
+  const [, setSelectedOperandInDropdown] = useRecoilScopedState(
+    selectedOperandInDropdownScopedState,
+    context,
+  );
+
+  React.useEffect(() => {
+    setFilterDefinitionUsedInDropdown(availableFilter);
+    const defaultOperand = getOperandsForFilterType(availableFilter?.type)[0];
+    setSelectedOperandInDropdown(defaultOperand);
+  }, [
+    availableFilter,
+    setFilterDefinitionUsedInDropdown,
+    setSelectedOperandInDropdown,
+  ]);
+
+  const setHotkeyScope = useSetHotkeyScope();
+
+  function handleIsUnfoldedChange(newIsUnfolded: boolean) {
+    if (newIsUnfolded) {
+      setHotkeyScope(HotkeyScope);
+      setIsUnfolded(true);
+    } else {
+      if (filterDefinitionUsedInDropdown?.type === 'entity') {
+        setHotkeyScope(HotkeyScope);
+      }
+      setIsUnfolded(false);
+      setFilterDropdownSearchInput('');
+    }
+  }
+
+  return (
+    <StyledDropdownButtonContainer>
+      <StyledDropdownButton
+        isUnfolded={isUnfolded}
+        onClick={() => handleIsUnfoldedChange(!isUnfolded)}
+      >
+        {filters[0]?.displayValue}
+      </StyledDropdownButton>
+      {isUnfolded && (
+        <DropdownMenuContainer onClose={() => handleIsUnfoldedChange(false)}>
+          <FilterDropdownEntitySearchInput context={context} />
+          <FilterDropdownEntitySelect context={context} />
+        </DropdownMenuContainer>
+      )}
+    </StyledDropdownButtonContainer>
+  );
+}
